requirements.txt
kafka-python
pytest
python-dateutil